continue migrating to exposed

+++ scraper/src/main/kotlin/com/nixops/scraper/tum_api/nat/model/NatProgram.kt
@@ -5,7 +5,7 @@ import com.fasterxml.jackson.annotation.JsonProperty
 
 @JsonIgnoreProperties(ignoreUnknown = true)
 data class NatProgram(
-    @JsonProperty("study_id") val studyId: Int,
+    @JsonProperty("study_id") val studyId: Long,
     @JsonProperty("org_id") val orgId: Int,
     @JsonProperty("school") val school: NatSchool,
     @JsonProperty("spo_version") val spoVersion: String,
